requirements/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
PyYAML
flask-bcrypt